Improve type checks in SaveObject.ts
